backend/requirements.txt
fastapi
uvicorn
yfinance
google-generativeai
python-dotenv
newsapi-python
pandas
matplotlib
requests
supabase
scikit-learn
ta

